feat: implement light mode and theme toggle functionality
- Added light mode CSS variables to globals.css for improved UI flexibility.
- Introduced ThemeProvider context to manage theme state across the application.
- Integrated ThemeToggle component for users to switch between light and dark modes.
- Updated layout and components to support theme changes, ensuring consistent styling.
- Enhanced GraphTemplateBuilder and GraphVisualization components to reflect theme changes.

These changes enhance user experience by providing a customizable interface that adapts to user preferences.

auth: @nk-ag
