+++ dashboard/src/components/GraphVisualization.tsx
@@ -34,10 +34,12 @@ import {
 import { Card, CardHeader, CardTitle, CardDescription, CardContent } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
+import { NodeDetailsModal } from '@/components/NodeDetailsModal';
 
 interface GraphVisualizationProps {
   namespace: string;
   runId: string;
+  onGraphTemplateRequest?: (graphName: string) => void;
 }
 
 // Custom Node Component
@@ -108,7 +110,7 @@ const CustomNode: React.FC<{
         </div>
       </div>
       <div className="text-sm font-medium text-card-foreground mb-1">{data.label}</div>
-      <div className="text-xs text-muted-foreground">{data.identifier}</div>
+      <div className="text-xs text-foreground">{data.identifier}</div>
     </div>
   );
 };
@@ -119,7 +121,8 @@ const nodeTypes: NodeTypes = {
 
 export const GraphVisualization: React.FC<GraphVisualizationProps> = ({
   namespace,
-  runId
+  runId,
+  onGraphTemplateRequest
 }) => {
   const [graphData, setGraphData] = useState<GraphStructureResponse | null>(null);
   const [isLoading, setIsLoading] = useState(false);
@@ -136,12 +139,17 @@ export const GraphVisualization: React.FC<GraphVisualizationProps> = ({
     try {
       const data = await clientApiService.getGraphStructure(namespace, runId);
       setGraphData(data);
+      
+      // Request graph template details if callback is provided and graph name exists
+      if (onGraphTemplateRequest && data.graph_name) {
+        onGraphTemplateRequest(data.graph_name);
+      }
     } catch (err) {
       setError(err instanceof Error ? err.message : 'Failed to load graph structure');
     } finally {
       setIsLoading(false);
     }
-  }, [namespace, runId]);
+  }, [namespace, runId, onGraphTemplateRequest]);
 
   const loadNodeDetails = useCallback(async (nodeId: string, graphName: string) => {
     setIsLoadingNodeDetails(true);
@@ -378,21 +386,115 @@ export const GraphVisualization: React.FC<GraphVisualizationProps> = ({
       </div>
 
       {/* Execution Summary */}
-      <Card>
-        <CardHeader>
+      <Card className="bg-card border-border">
+        <CardHeader className="pb-4">
           <div className="flex items-center space-x-2">
             <BarChart3 className="w-5 h-5 text-primary" />
             <CardTitle>Execution Summary</CardTitle>
           </div>
         </CardHeader>
         <CardContent>
-          <div className="grid grid-cols-2 md:grid-cols-4 lg:grid-cols-6 gap-4">
-            {Object.entries(graphData.execution_summary).map(([status, count]) => (
-              <div key={status} className="text-center">
-                <div className="text-2xl font-bold text-foreground">{count}</div>
-                <div className="text-sm text-muted-foreground capitalize">{status.toLowerCase()}</div>
-              </div>
-            ))}
+          <div className="grid grid-cols-2 md:grid-cols-4 lg:grid-cols-8 gap-4">
+            {Object.entries(graphData.execution_summary).map(([status, count]) => {
+              const getStatusConfig = (status: string) => {
+                switch (status.toLowerCase()) {
+                  case 'success':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                      text: 'text-chart-2', // Green from chart-2
+                      icon: <CheckCircle className="w-4 h-4" />,
+                      count: 'text-chart-2'
+                    };
+                  case 'created':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                      text: 'text-chart-1', // Sky blue from chart-1
+                      icon: <Clock className="w-4 h-4" />,
+                      count: 'text-chart-1'
+                    };
+                  case 'queued':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                      text: 'text-chart-3', // Yellow from chart-3
+                      icon: <Loader2 className="w-4 h-4" />,
+                      count: 'text-chart-3'
+                    };
+                  case 'executed':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                      text: 'text-chart-5', // Purple from chart-5
+                      icon: <CheckCircle className="w-4 h-4" />,
+                      count: 'text-chart-5'
+                    };
+                  case 'errored':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                      text: 'text-destructive', // Red from destructive
+                      icon: <XCircle className="w-4 h-4" />,
+                      count: 'text-destructive'
+                    };
+                  case 'next_created_error':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                        text: 'text-chart-4', // Pink from chart-4
+                      icon: <AlertCircle className="w-4 h-4" />,
+                      count: 'text-chart-4'
+                    };
+                  case 'pruned':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                      text: 'text-muted-foreground',
+                      icon: <XCircle className="w-4 h-4" />,
+                      count: 'text-muted-foreground'
+                    };
+                  case 'retry_created':
+                    return {
+                      bg: 'bg-secondary',
+                      border: 'border-border',
+                      text: 'text-primary', // Sky blue from primary
+                      icon: <RefreshCw className="w-4 h-4" />,
+                      count: 'text-primary'
+                    };
+                  default:
+                    return {
+                      bg: 'bg-muted',
+                      border: 'border-border',
+                      text: 'text-muted-foreground',
+                      icon: <Clock className="w-4 h-4" />,
+                      count: 'text-foreground'
+                    };
+                }
+              };
+              
+              const config = getStatusConfig(status);
+              const displayName = status.toLowerCase().replace(/_/g, ' ').replace(/\b\w/g, l => l.toUpperCase());
+              
+              return (
+                <div 
+                  key={status} 
+                  className={`relative p-4 rounded-lg border transition-all duration-200 hover:shadow-md hover:scale-105 ${config.bg} ${config.border}`}
+                >
+                  <div className="flex flex-col items-center space-y-2">
+                    <div className={`${config.text}`}>
+                      {config.icon}
+                    </div>
+                    <div className={`text-2xl font-bold ${config.count}`}>
+                      {count}
+                    </div>
+                    <div className={`text-xs font-medium text-center leading-tight ${config.text}`}>
+                      {displayName}
+                    </div>
+                  </div>
+                </div>
+              );
+            })}
           </div>
         </CardContent>
       </Card>
@@ -435,215 +537,19 @@ export const GraphVisualization: React.FC<GraphVisualizationProps> = ({
       </Card>
 
       {/* Node Details Modal */}
-      {selectedNode && (
-        <div className="fixed inset-0 bg-background/80 backdrop-blur-sm flex items-center justify-center z-50">
-          <Card className="max-w-2xl w-full mx-4 max-h-[80vh] overflow-y-auto">
-            <CardHeader>
-              <div className="flex items-center justify-between">
-                <CardTitle>Node Details</CardTitle>
-                <Button
-                  onClick={() => {
-                    setSelectedNode(null);
-                    setSelectedNodeDetails(null);
-                    setNodeDetailsError(null);
-                  }}
-                  variant="ghost"
-                  size="icon"
-                >
-                  <XCircle className="w-5 h-5" />
-                </Button>
-              </div>
-            </CardHeader>
-            
-            <CardContent className="space-y-6">
-              {/* Loading State */}
-              {isLoadingNodeDetails && (
-                <div className="flex items-center justify-center py-8">
-                  <Loader2 className="w-6 h-6 animate-spin text-primary" />
-                  <span className="ml-2 text-muted-foreground">Loading node details...</span>
-                </div>
-              )}
-
-              {/* Error State */}
-              {nodeDetailsError && (
-                <div className="flex items-center p-3 bg-destructive/5 border border-destructive/20 rounded-md">
-                  <AlertCircle className="w-5 h-5 text-destructive" />
-                  <div className="ml-3">
-                    <p className="text-sm text-destructive">{nodeDetailsError}</p>
-                  </div>
-                </div>
-              )}
-
-              {/* Node Header - always show basic info */}
-              <div className="flex items-center space-x-3">
-                {(() => {
-                  const status = selectedNodeDetails?.status || selectedNode.status;
-                  switch (status) {
-                    case 'CREATED':
-                      return <Clock className="w-5 h-5 text-muted-foreground" />;
-                    case 'QUEUED':
-                      return <Loader2 className="w-5 h-5 text-secondary animate-spin" />;
-                    case 'EXECUTED':
-                    case 'SUCCESS':
-                      return <CheckCircle className="w-5 h-5 text-chart-2" />;
-                    case 'ERRORED':
-                    case 'TIMEDOUT':
-                    case 'CANCELLED':
-                      return <XCircle className="w-5 h-5 text-destructive" />;
-                    default:
-                      return <Clock className="w-5 h-5 text-muted-foreground" />;
-                  }
-                })()}
-                <div className="flex-1">
-                  <h4 className="font-medium text-foreground">{selectedNodeDetails?.node_name || selectedNode.node_name}</h4>
-                  <p className="text-sm text-muted-foreground">{selectedNodeDetails?.identifier || selectedNode.identifier}</p>
-                </div>
-                <Badge variant={(() => {
-                  const status = selectedNodeDetails?.status || selectedNode.status;
-                  switch (status) {
-                    case 'EXECUTED':
-                    case 'SUCCESS':
-                      return 'success' as const;
-                    case 'ERRORED':
-                    case 'TIMEDOUT':
-                    case 'CANCELLED':
-                      return 'destructive' as const;
-                    case 'QUEUED':
-                      return 'secondary' as const;
-                    default:
-                      return 'default' as const;
-                  }
-                })()}>
-                  {selectedNodeDetails?.status || selectedNode.status}
-                </Badge>
-              </div>
-
-              {/* Only show detailed sections if not loading and no error */}
-              {!isLoadingNodeDetails && !nodeDetailsError && (
-                <>
-                  <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                    <div>
-                      <h5 className="font-medium text-foreground mb-3">Node Information</h5>
-                      <div className="space-y-2 text-sm">
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">ID:</span>
-                          <span className="text-foreground font-mono">{selectedNodeDetails?.id || selectedNode.id}</span>
-                        </div>
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">Name:</span>
-                          <span className="text-foreground">{selectedNodeDetails?.node_name || selectedNode.node_name}</span>
-                        </div>
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">Identifier:</span>
-                          <span className="text-foreground font-mono">{selectedNodeDetails?.identifier || selectedNode.identifier}</span>
-                        </div>
-                        {selectedNodeDetails?.graph_name && (
-                          <div className="flex justify-between">
-                            <span className="text-muted-foreground">Graph:</span>
-                            <span className="text-foreground">{selectedNodeDetails.graph_name}</span>
-                          </div>
-                        )}
-                        {selectedNodeDetails?.run_id && (
-                          <div className="flex justify-between">
-                            <span className="text-muted-foreground">Run ID:</span>
-                            <span className="text-foreground font-mono">{selectedNodeDetails.run_id}</span>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                    <div>
-                      <h5 className="font-medium text-foreground mb-3">Status & Timestamps</h5>
-                      <div className="space-y-2 text-sm">
-                        <div className="flex justify-between">
-                          <span className="text-muted-foreground">Current Status:</span>
-                          <span className="text-foreground">{selectedNodeDetails?.status || selectedNode.status}</span>
-                        </div>
-                        {selectedNodeDetails?.created_at && (
-                          <div className="flex justify-between">
-                            <span className="text-muted-foreground">Created:</span>
-                            <span className="text-foreground text-xs">{new Date(selectedNodeDetails.created_at).toLocaleString()}</span>
-                          </div>
-                        )}
-                        {selectedNodeDetails?.updated_at && (
-                          <div className="flex justify-between">
-                            <span className="text-muted-foreground">Updated:</span>
-                            <span className="text-foreground text-xs">{new Date(selectedNodeDetails.updated_at).toLocaleString()}</span>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-
-                  {/* Error Section */}
-                  {(selectedNodeDetails?.error || selectedNode.error) && (
-                    <div>
-                      <h5 className="font-medium text-destructive mb-3">Error</h5>
-                      <div className="text-sm text-destructive bg-destructive/5 border border-destructive/20 p-3 rounded-md">
-                        {selectedNodeDetails?.error || selectedNode.error}
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Parent Nodes Section */}
-                  {selectedNodeDetails?.parents && Object.keys(selectedNodeDetails.parents).length > 0 && (
-                    <div>
-                      <h5 className="font-medium text-foreground mb-3">Parent Nodes</h5>
-                      <div className="bg-muted/20 border border-border rounded-md p-3">
-                        <div className="space-y-1">
-                          {Object.entries(selectedNodeDetails.parents).map(([identifier, parentId]) => (
-                            <div key={identifier} className="flex justify-between text-sm">
-                              <span className="text-muted-foreground">{identifier}:</span>
-                              <span className="text-foreground font-mono">{parentId}</span>
-                            </div>
-                          ))}
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Inputs Section */}
-                  <div>
-                    <h5 className="font-medium text-foreground mb-3">Inputs</h5>
-                    <div className="bg-muted/20 border border-border rounded-md p-3">
-                      {(() => {
-                        const inputs = selectedNodeDetails?.inputs || selectedNode.inputs || {};
-                        return Object.keys(inputs).length > 0 ? (
-                          <pre className="text-sm text-foreground whitespace-pre-wrap font-mono">
-                            {JSON.stringify(inputs, null, 2)}
-                          </pre>
-                        ) : (
-                          <p className="text-sm text-muted-foreground italic">No inputs</p>
-                        );
-                      })()}
-                    </div>
-                  </div>
-
-                  {/* Outputs Section */}
-                  <div>
-                    <h5 className="font-medium text-foreground mb-3">Outputs</h5>
-                    <div className="bg-muted/20 border border-border rounded-md p-3">
-                      {(() => {
-                        const outputs = selectedNodeDetails?.outputs || selectedNode.outputs || {};
-                        return Object.keys(outputs).length > 0 ? (
-                          <pre className="text-sm text-foreground whitespace-pre-wrap font-mono">
-                            {JSON.stringify(outputs, null, 2)}
-                          </pre>
-                        ) : (
-                          <p className="text-sm text-muted-foreground italic">No outputs</p>
-                        );
-                      })()}
-                    </div>
-                  </div>
-                </>
-              )}
-
-              <div className="text-xs text-muted-foreground pt-4 border-t border-border">
-                Node ID: {selectedNodeDetails?.id || selectedNode.id}
-              </div>
-            </CardContent>
-          </Card>
-        </div>
-      )}
+      <NodeDetailsModal
+        selectedNode={selectedNode}
+        selectedNodeDetails={selectedNodeDetails}
+        isLoadingNodeDetails={isLoadingNodeDetails}
+        nodeDetailsError={nodeDetailsError}
+        namespace={namespace}
+        onClose={() => {
+          setSelectedNode(null);
+          setSelectedNodeDetails(null);
+          setNodeDetailsError(null);
+        }}
+        onRefreshGraph={loadGraphStructure}
+      />
     </div>
   );
 };
