+++ dashboard/src/components/ThemeToggle.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import React from 'react';
+import { Sun, Moon } from 'lucide-react';
+import { Button } from '@/components/ui/button';
+import { useTheme } from '@/contexts/ThemeContext';
+
+export const ThemeToggle: React.FC = () => {
+  const { theme, toggleTheme } = useTheme();
+
+  return (
+    <Button
+      variant="ghost"
+      size="icon"
+      onClick={toggleTheme}
+      className="relative"
+      aria-label={`Switch to ${theme === 'light' ? 'dark' : 'light'} mode`}
+    >
+      <Sun 
+        className={`h-[1.2rem] w-[1.2rem] transition-all ${
+          theme === 'light' ? 'rotate-0 scale-100' : 'rotate-90 scale-0'
+        }`} 
+      />
+      <Moon 
+        className={`absolute h-[1.2rem] w-[1.2rem] transition-all ${
+          theme === 'dark' ? 'rotate-0 scale-100' : 'rotate-90 scale-0'
+        }`} 
+      />
+    </Button>
+  );
+};
