+++ dashboard/src/services/clientApi.ts
@@ -30,6 +30,21 @@ export class ClientApiService {
     return response.json();
   }
 
+  // Manual Retry State
+  async manualRetryState(namespace: string, stateId: string, fanoutId: string) {
+    const response = await fetch(`/api/manual-retry-state?namespace=${encodeURIComponent(namespace)}&stateId=${encodeURIComponent(stateId)}`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ fanout_id: fanoutId }),
+    });
+    if (!response.ok) {
+      throw new Error(`Failed to retry state: ${response.statusText}`);
+    }
+    return response.json();
+  }
+
   // Namespace Overview
   async getNamespaceOverview(namespace: string) {
     const response = await fetch(`/api/namespace-overview?namespace=${encodeURIComponent(namespace)}`);
