+++ dashboard/src/types/state-manager.ts
@@ -245,3 +245,9 @@ export interface RunsResponse {
   size: number;
   runs: RunListItem[];
 }
+
+// Manual Retry Types
+export interface ManualRetryResponseModel {
+  id: string;
+  status: StateStatus;
+}
