+++ state-manager/tests/unit/controller/test_get_graph_structure.py
@@ -100,7 +100,9 @@ class TestGetGraphStructure:
             assert len(result.nodes) == 0
             assert len(result.edges) == 0
             assert len(result.root_states) == 0
-            assert result.execution_summary == {}
+            # All states should be initialized to 0
+            expected_summary = {status.value: 0 for status in StateStatusEnum}
+            assert result.execution_summary == expected_summary
 
     @pytest.mark.asyncio
     async def test_get_graph_structure_with_errors(self):
